Add. basic scripts for assuming role
This is a basic setup. It contains some manual changes
1. Creation of User manually in aws console. This user has permission to assume only one IAM role.
2. Creating IAM role which can be assumed by the newly created user.
3. This IAM role is given Admin permissions as of now but needs to be changed in future.

All this is needed because we want to assume the IAM role from local and do the operation without risking to expose actual access key ID and access key secret. It was only referred once when running aws configure. These creds are now used to assume the role and use the temporary creds to do all other operations.